Improve performance when adding servers to serverGroups (bsc#1158754)

+++ java/code/src/com/redhat/rhn/frontend/xmlrpc/system/test/ServerGroupHandlerTest.java
@@ -244,9 +244,9 @@ public class ServerGroupHandlerTest extends BaseHandlerTestCase {
 
     public void testRemoveNonExistentServer() throws Exception {
         ServerGroup group = handler.create(admin, NAME, DESCRIPTION);
-        List<Long> systems = new ArrayList<Long>();
+        List<Integer> systems = new ArrayList<Integer>();
         Server server1 = ServerFactoryTest.createTestServer(admin, true);
-        systems.add(server1.getId());
+        systems.add(server1.getId().intValue());
         try {
             handler.addOrRemoveSystems(admin, group.getName(), systems,
                     Boolean.FALSE);
